Port old events module to new events module
