+++ viste_reactive/src/old/mod.rs
@@ -1,2 +1,1 @@
-pub mod events;
 pub mod signals;
